requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain_ollama
langchain_groq
faiss-cpu
pdfplumber
python-dotenv